+++ src/ops/op_enum.h
@@ -287,6 +287,7 @@ typedef enum {
     E_OP_ES_STOP,
     E_OP_ES_TRIPLE,
     E_OP_ES_MAGIC,
+    E_OP_ES_CV,
     E_OP_OR_TRK,
     E_OP_OR_CLK,
     E_OP_OR_DIV,